requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
yt-dlp==2025.3.31
openai==1.68.0
anthropic==0.49.0
python-telegram-bot==21.10
python-dotenv==1.1.0
pytest==8.3.4
